requirements.txt
gspread>=5.12.0
pandas>=2.2.0
supabase>=2.7.4
python-dotenv>=1.0.0
oauth2client>=4.1.3
streamlit>=1.33.0
plotly>=5.18.0
